+++ NEWS.md
@@ -1,6 +1,8 @@
 # hubUtils 0.0.0.9014
 
-* Bug fix. Handle situation in `expand_model_out_val_grid()` when `required_vals_only = TRUE` yet required task ID columns are not consistent across modeling tasks. The function now pads missing task ID columns with `NA`s.
+* Added `coerce_to_character()` function for coercing all model output columns to character. This can be much faster than coercing to `coerce_to_hub_schema()`, especially for dates.
+* Added params `all_character` and `as_arrow_table` to `expand_model_out_val_grid()` to allow for returning all character columns and/or an arrow data table respectively.
+* Bug fix. Handle situation in `expand_model_out_val_grid()` when `required_vals_only = TRUE` yet required task ID columns are not consistent across modeling tasks. The function now pads missing task ID column values with `NA`s.
 
 # hubUtils 0.0.0.9013